+++ Editor/Headers/Editor.hpp
@@ -9,7 +9,7 @@ public:
 	float fps;
 	void update();
 	Renderer* getRenderer();
-	Editor(HWND &wndHandle, int width, int height, float fps);
+	Editor(HWND &wndHandle, float width, float height, float fps);
 	~Editor();
 private:
 	std::chrono::time_point<std::chrono::system_clock> lastTime;
